made some adjustments to the sign up pages
