+++ src/DoctorSignup/SignUpTop.jsx
@@ -1,0 +1,17 @@
+
+const SignUpTop = () => {
+    return (
+        <div className='w-full bg-white rounded-md'>
+            <div className='p-8 bg-white rounded-md'>
+                <h2 className='text-2xl font-bold mb-6 text-center text-violet-950'>
+                    Doctor Signup
+                </h2>
+                <p className='text-center mb-8 text-violet-950'>
+                    Input correct details to signup
+                </p>
+            </div>
+        </div>
+            );
+};
+
+export default SignUpTop;
