+++ src/DoctorSignup/VerificationSuccessful.jsx
@@ -2,581 +2,9 @@ import React from 'react'
 
 const VerificationSuccessful = () => {
   return (
-    <div className='flex items-center justify-center min-h-screen bg-white'>
+    <div className='flex items-center justify-center bg-white'>
       <div className='bg-gray-50 p-8 rounded-lg shadow-lg text-center max-w-md w-full'>
-        {/* <svg
-          className='h-40 w-60 mx-auto mb-4 text-blue-600'
-          width='963'
-          height='685'
-          viewBox='0 0 963 685'
-          fill='none'
-          xmlns='http://www.w3.org/2000/svg'
-        >
-          <rect width='963' height='685' rx='10' fill='white' />
-          <path
-            d='M345.359 500.844L351.344 482.25H355.672L347.516 505H344.594L345.359 500.844ZM339.828 482.25L345.766 500.844L346.578 505H343.641L335.516 482.25H339.828ZM364.609 505.312C363.359 505.312 362.229 505.109 361.219 504.703C360.219 504.286 359.365 503.708 358.656 502.969C357.958 502.229 357.422 501.359 357.047 500.359C356.672 499.359 356.484 498.281 356.484 497.125V496.5C356.484 495.177 356.677 493.979 357.062 492.906C357.448 491.833 357.984 490.917 358.672 490.156C359.359 489.385 360.172 488.797 361.109 488.391C362.047 487.984 363.062 487.781 364.156 487.781C365.365 487.781 366.422 487.984 367.328 488.391C368.234 488.797 368.984 489.37 369.578 490.109C370.182 490.839 370.63 491.708 370.922 492.719C371.224 493.729 371.375 494.844 371.375 496.062V497.672H358.312V494.969H367.656V494.672C367.635 493.995 367.5 493.359 367.25 492.766C367.01 492.172 366.641 491.693 366.141 491.328C365.641 490.964 364.974 490.781 364.141 490.781C363.516 490.781 362.958 490.917 362.469 491.188C361.99 491.448 361.589 491.828 361.266 492.328C360.943 492.828 360.693 493.432 360.516 494.141C360.349 494.839 360.266 495.625 360.266 496.5V497.125C360.266 497.865 360.365 498.552 360.562 499.188C360.771 499.812 361.073 500.359 361.469 500.828C361.865 501.297 362.344 501.667 362.906 501.938C363.469 502.198 364.109 502.328 364.828 502.328C365.734 502.328 366.542 502.146 367.25 501.781C367.958 501.417 368.573 500.901 369.094 500.234L371.078 502.156C370.714 502.688 370.24 503.198 369.656 503.688C369.073 504.167 368.359 504.557 367.516 504.859C366.682 505.161 365.714 505.312 364.609 505.312ZM378.094 491.312V505H374.328V488.094H377.922L378.094 491.312ZM383.266 487.984L383.234 491.484C383.005 491.443 382.755 491.411 382.484 491.391C382.224 491.37 381.964 491.359 381.703 491.359C381.057 491.359 380.49 491.453 380 491.641C379.51 491.818 379.099 492.078 378.766 492.422C378.443 492.755 378.193 493.161 378.016 493.641C377.839 494.12 377.734 494.656 377.703 495.25L376.844 495.312C376.844 494.25 376.948 493.266 377.156 492.359C377.365 491.453 377.677 490.656 378.094 489.969C378.521 489.281 379.052 488.745 379.688 488.359C380.333 487.974 381.078 487.781 381.922 487.781C382.151 487.781 382.396 487.802 382.656 487.844C382.927 487.885 383.13 487.932 383.266 487.984ZM389.656 488.094V505H385.875V488.094H389.656ZM385.625 483.656C385.625 483.083 385.812 482.609 386.188 482.234C386.573 481.849 387.104 481.656 387.781 481.656C388.448 481.656 388.974 481.849 389.359 482.234C389.745 482.609 389.938 483.083 389.938 483.656C389.938 484.219 389.745 484.688 389.359 485.062C388.974 485.438 388.448 485.625 387.781 485.625C387.104 485.625 386.573 485.438 386.188 485.062C385.812 484.688 385.625 484.219 385.625 483.656ZM398.625 505H394.844V487C394.844 485.635 395.115 484.484 395.656 483.547C396.208 482.609 396.995 481.896 398.016 481.406C399.047 480.917 400.271 480.672 401.688 480.672C402.521 480.672 403.323 480.76 404.094 480.938C404.875 481.115 405.682 481.339 406.516 481.609L405.938 484.672C405.406 484.495 404.818 484.328 404.172 484.172C403.526 484.005 402.781 483.922 401.938 483.922C400.802 483.922 399.964 484.182 399.422 484.703C398.891 485.214 398.625 485.979 398.625 487V505ZM402.016 488.094V490.844H392.25V488.094H402.016ZM408.297 488.094V505H404.531V488.094H408.297ZM419.406 502.312C420.021 502.312 420.573 502.193 421.062 501.953C421.562 501.703 421.964 501.359 422.266 500.922C422.578 500.484 422.75 499.979 422.781 499.406H426.328C426.307 500.5 425.984 501.495 425.359 502.391C424.734 503.286 423.906 504 422.875 504.531C421.844 505.052 420.703 505.312 419.453 505.312C418.161 505.312 417.036 505.094 416.078 504.656C415.12 504.208 414.323 503.594 413.688 502.812C413.052 502.031 412.573 501.13 412.25 500.109C411.938 499.089 411.781 497.995 411.781 496.828V496.281C411.781 495.115 411.938 494.021 412.25 493C412.573 491.969 413.052 491.062 413.688 490.281C414.323 489.5 415.12 488.891 416.078 488.453C417.036 488.005 418.156 487.781 419.438 487.781C420.792 487.781 421.979 488.052 423 488.594C424.021 489.125 424.823 489.87 425.406 490.828C426 491.776 426.307 492.88 426.328 494.141H422.781C422.75 493.516 422.594 492.953 422.312 492.453C422.042 491.943 421.656 491.536 421.156 491.234C420.667 490.932 420.078 490.781 419.391 490.781C418.63 490.781 418 490.938 417.5 491.25C417 491.552 416.609 491.969 416.328 492.5C416.047 493.021 415.844 493.609 415.719 494.266C415.604 494.911 415.547 495.583 415.547 496.281V496.828C415.547 497.526 415.604 498.203 415.719 498.859C415.833 499.516 416.031 500.104 416.312 500.625C416.604 501.135 417 501.547 417.5 501.859C418 502.161 418.635 502.312 419.406 502.312ZM438.75 501.609V493.547C438.75 492.943 438.641 492.422 438.422 491.984C438.203 491.547 437.87 491.208 437.422 490.969C436.984 490.729 436.432 490.609 435.766 490.609C435.151 490.609 434.62 490.714 434.172 490.922C433.724 491.13 433.375 491.411 433.125 491.766C432.875 492.12 432.75 492.521 432.75 492.969H429C429 492.302 429.161 491.656 429.484 491.031C429.807 490.406 430.276 489.849 430.891 489.359C431.505 488.87 432.24 488.484 433.094 488.203C433.948 487.922 434.906 487.781 435.969 487.781C437.24 487.781 438.365 487.995 439.344 488.422C440.333 488.849 441.109 489.495 441.672 490.359C442.245 491.214 442.531 492.286 442.531 493.578V501.094C442.531 501.865 442.583 502.557 442.688 503.172C442.802 503.776 442.964 504.302 443.172 504.75V505H439.312C439.135 504.594 438.995 504.078 438.891 503.453C438.797 502.818 438.75 502.203 438.75 501.609ZM439.297 494.719L439.328 497.047H436.625C435.927 497.047 435.312 497.115 434.781 497.25C434.25 497.375 433.807 497.562 433.453 497.812C433.099 498.062 432.833 498.365 432.656 498.719C432.479 499.073 432.391 499.474 432.391 499.922C432.391 500.37 432.495 500.781 432.703 501.156C432.911 501.521 433.214 501.807 433.609 502.016C434.016 502.224 434.505 502.328 435.078 502.328C435.849 502.328 436.521 502.172 437.094 501.859C437.677 501.536 438.135 501.146 438.469 500.688C438.802 500.219 438.979 499.776 439 499.359L440.219 501.031C440.094 501.458 439.88 501.917 439.578 502.406C439.276 502.896 438.88 503.365 438.391 503.812C437.911 504.25 437.333 504.609 436.656 504.891C435.99 505.172 435.219 505.312 434.344 505.312C433.24 505.312 432.255 505.094 431.391 504.656C430.526 504.208 429.849 503.609 429.359 502.859C428.87 502.099 428.625 501.24 428.625 500.281C428.625 499.385 428.792 498.594 429.125 497.906C429.469 497.208 429.969 496.625 430.625 496.156C431.292 495.688 432.104 495.333 433.062 495.094C434.021 494.844 435.115 494.719 436.344 494.719H439.297ZM454.219 488.094V490.844H444.688V488.094H454.219ZM447.438 483.953H451.203V500.328C451.203 500.849 451.276 501.25 451.422 501.531C451.578 501.802 451.792 501.984 452.062 502.078C452.333 502.172 452.651 502.219 453.016 502.219C453.276 502.219 453.526 502.203 453.766 502.172C454.005 502.141 454.198 502.109 454.344 502.078L454.359 504.953C454.047 505.047 453.682 505.13 453.266 505.203C452.859 505.276 452.391 505.312 451.859 505.312C450.995 505.312 450.229 505.161 449.562 504.859C448.896 504.547 448.375 504.042 448 503.344C447.625 502.646 447.438 501.719 447.438 500.562V483.953ZM461.156 488.094V505H457.375V488.094H461.156ZM457.125 483.656C457.125 483.083 457.312 482.609 457.688 482.234C458.073 481.849 458.604 481.656 459.281 481.656C459.948 481.656 460.474 481.849 460.859 482.234C461.245 482.609 461.438 483.083 461.438 483.656C461.438 484.219 461.245 484.688 460.859 485.062C460.474 485.438 459.948 485.625 459.281 485.625C458.604 485.625 458.073 485.438 457.688 485.062C457.312 484.688 457.125 484.219 457.125 483.656ZM464.562 496.734V496.375C464.562 495.156 464.74 494.026 465.094 492.984C465.448 491.932 465.958 491.021 466.625 490.25C467.302 489.469 468.125 488.865 469.094 488.438C470.073 488 471.177 487.781 472.406 487.781C473.646 487.781 474.75 488 475.719 488.438C476.698 488.865 477.526 489.469 478.203 490.25C478.88 491.021 479.396 491.932 479.75 492.984C480.104 494.026 480.281 495.156 480.281 496.375V496.734C480.281 497.953 480.104 499.083 479.75 500.125C479.396 501.167 478.88 502.078 478.203 502.859C477.526 503.63 476.703 504.234 475.734 504.672C474.766 505.099 473.667 505.312 472.438 505.312C471.198 505.312 470.089 505.099 469.109 504.672C468.141 504.234 467.318 503.63 466.641 502.859C465.964 502.078 465.448 501.167 465.094 500.125C464.74 499.083 464.562 497.953 464.562 496.734ZM468.328 496.375V496.734C468.328 497.495 468.406 498.214 468.562 498.891C468.719 499.568 468.964 500.161 469.297 500.672C469.63 501.182 470.057 501.583 470.578 501.875C471.099 502.167 471.719 502.312 472.438 502.312C473.135 502.312 473.74 502.167 474.25 501.875C474.771 501.583 475.198 501.182 475.531 500.672C475.865 500.161 476.109 499.568 476.266 498.891C476.432 498.214 476.516 497.495 476.516 496.734V496.375C476.516 495.625 476.432 494.917 476.266 494.25C476.109 493.573 475.859 492.974 475.516 492.453C475.182 491.932 474.755 491.526 474.234 491.234C473.724 490.932 473.115 490.781 472.406 490.781C471.698 490.781 471.083 490.932 470.562 491.234C470.052 491.526 469.63 491.932 469.297 492.453C468.964 492.974 468.719 493.573 468.562 494.25C468.406 494.917 468.328 495.625 468.328 496.375ZM487.172 491.703V505H483.406V488.094H486.953L487.172 491.703ZM486.5 495.922L485.281 495.906C485.292 494.708 485.458 493.609 485.781 492.609C486.115 491.609 486.573 490.75 487.156 490.031C487.75 489.312 488.458 488.76 489.281 488.375C490.104 487.979 491.021 487.781 492.031 487.781C492.844 487.781 493.578 487.896 494.234 488.125C494.901 488.344 495.469 488.703 495.938 489.203C496.417 489.703 496.781 490.354 497.031 491.156C497.281 491.948 497.406 492.922 497.406 494.078V505H493.625V494.062C493.625 493.25 493.505 492.609 493.266 492.141C493.036 491.661 492.698 491.323 492.25 491.125C491.812 490.917 491.266 490.812 490.609 490.812C489.964 490.812 489.385 490.948 488.875 491.219C488.365 491.49 487.932 491.859 487.578 492.328C487.234 492.797 486.969 493.339 486.781 493.953C486.594 494.568 486.5 495.224 486.5 495.922ZM521.562 499.141C521.562 498.672 521.49 498.255 521.344 497.891C521.208 497.526 520.964 497.193 520.609 496.891C520.255 496.589 519.755 496.297 519.109 496.016C518.474 495.724 517.661 495.427 516.672 495.125C515.589 494.792 514.589 494.422 513.672 494.016C512.766 493.599 511.974 493.12 511.297 492.578C510.62 492.026 510.094 491.396 509.719 490.688C509.344 489.969 509.156 489.141 509.156 488.203C509.156 487.276 509.349 486.432 509.734 485.672C510.13 484.911 510.688 484.255 511.406 483.703C512.135 483.141 512.995 482.708 513.984 482.406C514.974 482.094 516.068 481.938 517.266 481.938C518.953 481.938 520.406 482.25 521.625 482.875C522.854 483.5 523.797 484.339 524.453 485.391C525.12 486.443 525.453 487.604 525.453 488.875H521.562C521.562 488.125 521.401 487.464 521.078 486.891C520.766 486.307 520.286 485.849 519.641 485.516C519.005 485.182 518.198 485.016 517.219 485.016C516.292 485.016 515.521 485.156 514.906 485.438C514.292 485.719 513.833 486.099 513.531 486.578C513.229 487.057 513.078 487.599 513.078 488.203C513.078 488.63 513.177 489.021 513.375 489.375C513.573 489.719 513.875 490.042 514.281 490.344C514.688 490.635 515.198 490.911 515.812 491.172C516.427 491.432 517.151 491.682 517.984 491.922C519.245 492.297 520.344 492.714 521.281 493.172C522.219 493.62 523 494.13 523.625 494.703C524.25 495.276 524.719 495.927 525.031 496.656C525.344 497.375 525.5 498.193 525.5 499.109C525.5 500.068 525.307 500.932 524.922 501.703C524.536 502.464 523.984 503.115 523.266 503.656C522.557 504.188 521.703 504.599 520.703 504.891C519.714 505.172 518.609 505.312 517.391 505.312C516.297 505.312 515.219 505.167 514.156 504.875C513.104 504.583 512.146 504.141 511.281 503.547C510.417 502.943 509.729 502.193 509.219 501.297C508.708 500.391 508.453 499.333 508.453 498.125H512.375C512.375 498.865 512.5 499.495 512.75 500.016C513.01 500.536 513.37 500.964 513.828 501.297C514.286 501.62 514.818 501.859 515.422 502.016C516.036 502.172 516.693 502.25 517.391 502.25C518.307 502.25 519.073 502.12 519.688 501.859C520.312 501.599 520.781 501.234 521.094 500.766C521.406 500.297 521.562 499.755 521.562 499.141ZM538.766 501.016V488.094H542.547V505H538.984L538.766 501.016ZM539.297 497.5L540.562 497.469C540.562 498.604 540.438 499.651 540.188 500.609C539.938 501.557 539.552 502.385 539.031 503.094C538.51 503.792 537.844 504.339 537.031 504.734C536.219 505.12 535.245 505.312 534.109 505.312C533.286 505.312 532.531 505.193 531.844 504.953C531.156 504.714 530.562 504.344 530.062 503.844C529.573 503.344 529.193 502.693 528.922 501.891C528.651 501.089 528.516 500.13 528.516 499.016V488.094H532.281V499.047C532.281 499.661 532.354 500.177 532.5 500.594C532.646 501 532.844 501.328 533.094 501.578C533.344 501.828 533.635 502.005 533.969 502.109C534.302 502.214 534.656 502.266 535.031 502.266C536.104 502.266 536.948 502.057 537.562 501.641C538.188 501.214 538.63 500.641 538.891 499.922C539.161 499.203 539.297 498.396 539.297 497.5ZM553.312 502.312C553.927 502.312 554.479 502.193 554.969 501.953C555.469 501.703 555.87 501.359 556.172 500.922C556.484 500.484 556.656 499.979 556.688 499.406H560.234C560.214 500.5 559.891 501.495 559.266 502.391C558.641 503.286 557.812 504 556.781 504.531C555.75 505.052 554.609 505.312 553.359 505.312C552.068 505.312 550.943 505.094 549.984 504.656C549.026 504.208 548.229 503.594 547.594 502.812C546.958 502.031 546.479 501.13 546.156 500.109C545.844 499.089 545.688 497.995 545.688 496.828V496.281C545.688 495.115 545.844 494.021 546.156 493C546.479 491.969 546.958 491.062 547.594 490.281C548.229 489.5 549.026 488.891 549.984 488.453C550.943 488.005 552.062 487.781 553.344 487.781C554.698 487.781 555.885 488.052 556.906 488.594C557.927 489.125 558.729 489.87 559.312 490.828C559.906 491.776 560.214 492.88 560.234 494.141H556.688C556.656 493.516 556.5 492.953 556.219 492.453C555.948 491.943 555.562 491.536 555.062 491.234C554.573 490.932 553.984 490.781 553.297 490.781C552.536 490.781 551.906 490.938 551.406 491.25C550.906 491.552 550.516 491.969 550.234 492.5C549.953 493.021 549.75 493.609 549.625 494.266C549.51 494.911 549.453 495.583 549.453 496.281V496.828C549.453 497.526 549.51 498.203 549.625 498.859C549.74 499.516 549.938 500.104 550.219 500.625C550.51 501.135 550.906 501.547 551.406 501.859C551.906 502.161 552.542 502.312 553.312 502.312ZM570.031 502.312C570.646 502.312 571.198 502.193 571.688 501.953C572.188 501.703 572.589 501.359 572.891 500.922C573.203 500.484 573.375 499.979 573.406 499.406H576.953C576.932 500.5 576.609 501.495 575.984 502.391C575.359 503.286 574.531 504 573.5 504.531C572.469 505.052 571.328 505.312 570.078 505.312C568.786 505.312 567.661 505.094 566.703 504.656C565.745 504.208 564.948 503.594 564.312 502.812C563.677 502.031 563.198 501.13 562.875 500.109C562.562 499.089 562.406 497.995 562.406 496.828V496.281C562.406 495.115 562.562 494.021 562.875 493C563.198 491.969 563.677 491.062 564.312 490.281C564.948 489.5 565.745 488.891 566.703 488.453C567.661 488.005 568.781 487.781 570.062 487.781C571.417 487.781 572.604 488.052 573.625 488.594C574.646 489.125 575.448 489.87 576.031 490.828C576.625 491.776 576.932 492.88 576.953 494.141H573.406C573.375 493.516 573.219 492.953 572.938 492.453C572.667 491.943 572.281 491.536 571.781 491.234C571.292 490.932 570.703 490.781 570.016 490.781C569.255 490.781 568.625 490.938 568.125 491.25C567.625 491.552 567.234 491.969 566.953 492.5C566.672 493.021 566.469 493.609 566.344 494.266C566.229 494.911 566.172 495.583 566.172 496.281V496.828C566.172 497.526 566.229 498.203 566.344 498.859C566.458 499.516 566.656 500.104 566.938 500.625C567.229 501.135 567.625 501.547 568.125 501.859C568.625 502.161 569.26 502.312 570.031 502.312ZM587.297 505.312C586.047 505.312 584.917 505.109 583.906 504.703C582.906 504.286 582.052 503.708 581.344 502.969C580.646 502.229 580.109 501.359 579.734 500.359C579.359 499.359 579.172 498.281 579.172 497.125V496.5C579.172 495.177 579.365 493.979 579.75 492.906C580.135 491.833 580.672 490.917 581.359 490.156C582.047 489.385 582.859 488.797 583.797 488.391C584.734 487.984 585.75 487.781 586.844 487.781C588.052 487.781 589.109 487.984 590.016 488.391C590.922 488.797 591.672 489.37 592.266 490.109C592.87 490.839 593.318 491.708 593.609 492.719C593.911 493.729 594.062 494.844 594.062 496.062V497.672H581V494.969H590.344V494.672C590.323 493.995 590.188 493.359 589.938 492.766C589.698 492.172 589.328 491.693 588.828 491.328C588.328 490.964 587.661 490.781 586.828 490.781C586.203 490.781 585.646 490.917 585.156 491.188C584.677 491.448 584.276 491.828 583.953 492.328C583.63 492.828 583.38 493.432 583.203 494.141C583.036 494.839 582.953 495.625 582.953 496.5V497.125C582.953 497.865 583.052 498.552 583.25 499.188C583.458 499.812 583.76 500.359 584.156 500.828C584.552 501.297 585.031 501.667 585.594 501.938C586.156 502.198 586.797 502.328 587.516 502.328C588.422 502.328 589.229 502.146 589.938 501.781C590.646 501.417 591.26 500.901 591.781 500.234L593.766 502.156C593.401 502.688 592.927 503.198 592.344 503.688C591.76 504.167 591.047 504.557 590.203 504.859C589.37 505.161 588.401 505.312 587.297 505.312ZM606.484 500.422C606.484 500.047 606.391 499.708 606.203 499.406C606.016 499.094 605.656 498.812 605.125 498.562C604.604 498.312 603.833 498.083 602.812 497.875C601.917 497.677 601.094 497.443 600.344 497.172C599.604 496.891 598.969 496.552 598.438 496.156C597.906 495.76 597.495 495.292 597.203 494.75C596.911 494.208 596.766 493.583 596.766 492.875C596.766 492.188 596.917 491.536 597.219 490.922C597.521 490.307 597.953 489.766 598.516 489.297C599.078 488.828 599.76 488.458 600.562 488.188C601.375 487.917 602.281 487.781 603.281 487.781C604.698 487.781 605.911 488.021 606.922 488.5C607.943 488.969 608.724 489.609 609.266 490.422C609.807 491.224 610.078 492.13 610.078 493.141H606.312C606.312 492.693 606.198 492.276 605.969 491.891C605.75 491.495 605.417 491.177 604.969 490.938C604.521 490.688 603.958 490.562 603.281 490.562C602.635 490.562 602.099 490.667 601.672 490.875C601.255 491.073 600.943 491.333 600.734 491.656C600.536 491.979 600.438 492.333 600.438 492.719C600.438 493 600.49 493.255 600.594 493.484C600.708 493.703 600.896 493.906 601.156 494.094C601.417 494.271 601.771 494.438 602.219 494.594C602.677 494.75 603.25 494.901 603.938 495.047C605.229 495.318 606.339 495.667 607.266 496.094C608.203 496.51 608.922 497.052 609.422 497.719C609.922 498.375 610.172 499.208 610.172 500.219C610.172 500.969 610.01 501.656 609.688 502.281C609.375 502.896 608.917 503.432 608.312 503.891C607.708 504.339 606.984 504.688 606.141 504.938C605.307 505.188 604.37 505.312 603.328 505.312C601.797 505.312 600.5 505.042 599.438 504.5C598.375 503.948 597.568 503.245 597.016 502.391C596.474 501.526 596.203 500.63 596.203 499.703H599.844C599.885 500.401 600.078 500.958 600.422 501.375C600.776 501.781 601.214 502.078 601.734 502.266C602.266 502.443 602.812 502.531 603.375 502.531C604.052 502.531 604.62 502.443 605.078 502.266C605.536 502.078 605.885 501.828 606.125 501.516C606.365 501.193 606.484 500.828 606.484 500.422ZM622.953 500.422C622.953 500.047 622.859 499.708 622.672 499.406C622.484 499.094 622.125 498.812 621.594 498.562C621.073 498.312 620.302 498.083 619.281 497.875C618.385 497.677 617.562 497.443 616.812 497.172C616.073 496.891 615.438 496.552 614.906 496.156C614.375 495.76 613.964 495.292 613.672 494.75C613.38 494.208 613.234 493.583 613.234 492.875C613.234 492.188 613.385 491.536 613.688 490.922C613.99 490.307 614.422 489.766 614.984 489.297C615.547 488.828 616.229 488.458 617.031 488.188C617.844 487.917 618.75 487.781 619.75 487.781C621.167 487.781 622.38 488.021 623.391 488.5C624.411 488.969 625.193 489.609 625.734 490.422C626.276 491.224 626.547 492.13 626.547 493.141H622.781C622.781 492.693 622.667 492.276 622.438 491.891C622.219 491.495 621.885 491.177 621.438 490.938C620.99 490.688 620.427 490.562 619.75 490.562C619.104 490.562 618.568 490.667 618.141 490.875C617.724 491.073 617.411 491.333 617.203 491.656C617.005 491.979 616.906 492.333 616.906 492.719C616.906 493 616.958 493.255 617.062 493.484C617.177 493.703 617.365 493.906 617.625 494.094C617.885 494.271 618.24 494.438 618.688 494.594C619.146 494.75 619.719 494.901 620.406 495.047C621.698 495.318 622.807 495.667 623.734 496.094C624.672 496.51 625.391 497.052 625.891 497.719C626.391 498.375 626.641 499.208 626.641 500.219C626.641 500.969 626.479 501.656 626.156 502.281C625.844 502.896 625.385 503.432 624.781 503.891C624.177 504.339 623.453 504.688 622.609 504.938C621.776 505.188 620.839 505.312 619.797 505.312C618.266 505.312 616.969 505.042 615.906 504.5C614.844 503.948 614.036 503.245 613.484 502.391C612.943 501.526 612.672 500.63 612.672 499.703H616.312C616.354 500.401 616.547 500.958 616.891 501.375C617.245 501.781 617.682 502.078 618.203 502.266C618.734 502.443 619.281 502.531 619.844 502.531C620.521 502.531 621.089 502.443 621.547 502.266C622.005 502.078 622.354 501.828 622.594 501.516C622.833 501.193 622.953 500.828 622.953 500.422ZM635.031 505H631.266V486.469C631.266 485.208 631.5 484.151 631.969 483.297C632.448 482.432 633.13 481.781 634.016 481.344C634.901 480.896 635.948 480.672 637.156 480.672C637.531 480.672 637.901 480.698 638.266 480.75C638.63 480.792 638.984 480.859 639.328 480.953L639.234 483.859C639.026 483.807 638.797 483.771 638.547 483.75C638.307 483.729 638.047 483.719 637.766 483.719C637.193 483.719 636.698 483.828 636.281 484.047C635.875 484.255 635.562 484.562 635.344 484.969C635.135 485.375 635.031 485.875 635.031 486.469V505ZM638.516 488.094V490.844H628.672V488.094H638.516ZM651.422 501.016V488.094H655.203V505H651.641L651.422 501.016ZM651.953 497.5L653.219 497.469C653.219 498.604 653.094 499.651 652.844 500.609C652.594 501.557 652.208 502.385 651.688 503.094C651.167 503.792 650.5 504.339 649.688 504.734C648.875 505.12 647.901 505.312 646.766 505.312C645.943 505.312 645.188 505.193 644.5 504.953C643.812 504.714 643.219 504.344 642.719 503.844C642.229 503.344 641.849 502.693 641.578 501.891C641.307 501.089 641.172 500.13 641.172 499.016V488.094H644.938V499.047C644.938 499.661 645.01 500.177 645.156 500.594C645.302 501 645.5 501.328 645.75 501.578C646 501.828 646.292 502.005 646.625 502.109C646.958 502.214 647.312 502.266 647.688 502.266C648.76 502.266 649.604 502.057 650.219 501.641C650.844 501.214 651.286 500.641 651.547 499.922C651.818 499.203 651.953 498.396 651.953 497.5ZM663.094 481V505H659.312V481H663.094ZM671.562 482.25L671.188 498.297H667.969L667.578 482.25H671.562ZM667.469 503.219C667.469 502.646 667.656 502.167 668.031 501.781C668.417 501.385 668.948 501.188 669.625 501.188C670.292 501.188 670.818 501.385 671.203 501.781C671.589 502.167 671.781 502.646 671.781 503.219C671.781 503.771 671.589 504.245 671.203 504.641C670.818 505.026 670.292 505.219 669.625 505.219C668.948 505.219 668.417 505.026 668.031 504.641C667.656 504.245 667.469 503.771 667.469 503.219Z'
-            fill='#020E7C'
-          />
-          <circle cx='479.283' cy='257.026' r='83.009' fill='#136CFB' />
-          <g filter='url(#filter0_d_1_4116)'>
-            <path
-              d='M450.484 262.955L470.899 280.742L508.082 233.309'
-              stroke='white'
-              stroke-width='5'
-              stroke-linecap='round'
-              stroke-linejoin='round'
-            />
-          </g>
-          <rect
-            x='488.065'
-            y='125.337'
-            width='8.30228'
-            height='9.686'
-            rx='1'
-            transform='rotate(-10.9119 488.065 125.337)'
-            fill='url(#paint0_linear_1_4116)'
-          />
-          <rect
-            x='618.508'
-            y='332.32'
-            width='8.30228'
-            height='9.686'
-            rx='1'
-            transform='rotate(29.09 618.508 332.32)'
-            fill='url(#paint1_linear_1_4116)'
-          />
-          <rect
-            x='385.496'
-            y='396.5'
-            width='8.30228'
-            height='9.686'
-            rx='1'
-            transform='rotate(29.09 385.496 396.5)'
-            fill='url(#paint2_linear_1_4116)'
-          />
-          <rect
-            x='432.757'
-            y='152.546'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint3_linear_1_4116)'
-          />
-          <rect
-            x='612.781'
-            y='120.925'
-            width='7.05463'
-            height='8.23041'
-            rx='1'
-            transform='rotate(27 612.781 120.925)'
-            fill='url(#paint4_linear_1_4116)'
-          />
-          <rect
-            x='601.875'
-            y='157.066'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint5_linear_1_4116)'
-          />
-          <rect
-            x='590.573'
-            y='324.957'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint6_linear_1_4116)'
-          />
-          <rect
-            x='621.245'
-            y='260.384'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint7_linear_1_4116)'
-          />
-          <rect
-            x='664.833'
-            y='202.268'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint8_linear_1_4116)'
-          />
-          <rect
-            x='629.314'
-            y='305.584'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint9_linear_1_4116)'
-          />
-          <rect
-            x='550.435'
-            y='390.856'
-            width='3.8306'
-            height='4.46904'
-            rx='1'
-            fill='url(#paint10_linear_1_4116)'
-          />
-          <rect
-            x='388.586'
-            y='136.435'
-            width='5.46975'
-            height='6.38138'
-            rx='1'
-            transform='rotate(-22.65 388.586 136.435)'
-            fill='url(#paint11_linear_1_4116)'
-          />
-          <rect
-            x='284.689'
-            y='364.433'
-            width='5.46975'
-            height='6.38138'
-            rx='1'
-            transform='rotate(17.3473 284.689 364.433)'
-            fill='url(#paint12_linear_1_4116)'
-          />
-          <rect
-            x='371.748'
-            y='174.824'
-            width='2.07146'
-            height='2.4167'
-            rx='1'
-            transform='rotate(17.3473 371.748 174.824)'
-            fill='url(#paint13_linear_1_4116)'
-          />
-          <rect
-            x='583.221'
-            y='187.74'
-            width='2.07146'
-            height='2.4167'
-            rx='1'
-            transform='rotate(17.3473 583.221 187.74)'
-            fill='url(#paint14_linear_1_4116)'
-          />
-          <rect
-            x='668.781'
-            y='224.868'
-            width='2.07146'
-            height='2.4167'
-            rx='1'
-            transform='rotate(17.3473 668.781 224.868)'
-            fill='url(#paint15_linear_1_4116)'
-          />
-          <rect
-            x='388.884'
-            y='325.549'
-            width='2.07146'
-            height='2.4167'
-            rx='1'
-            transform='rotate(17.3473 388.884 325.549)'
-            fill='url(#paint16_linear_1_4116)'
-          />
-          <rect
-            x='313.63'
-            y='98.9502'
-            width='2.07146'
-            height='2.4167'
-            rx='1'
-            transform='rotate(17.3473 313.63 98.9502)'
-            fill='url(#paint17_linear_1_4116)'
-          />
-          <rect
-            x='328.804'
-            y='255.082'
-            width='3.93352'
-            height='4.58911'
-            rx='1'
-            transform='rotate(17.3473 328.804 255.082)'
-            fill='url(#paint18_linear_1_4116)'
-          />
-          <rect
-            x='333.879'
-            y='313.467'
-            width='3.93352'
-            height='4.58911'
-            rx='1'
-            transform='rotate(17.3473 333.879 313.467)'
-            fill='url(#paint19_linear_1_4116)'
-          />
-          <rect
-            x='449.863'
-            y='369.769'
-            width='3.93352'
-            height='4.58911'
-            rx='1'
-            transform='rotate(17.3473 449.863 369.769)'
-            fill='url(#paint20_linear_1_4116)'
-          />
-          <rect
-            x='516.6'
-            y='97.3374'
-            width='5.46975'
-            height='6.38138'
-            rx='1'
-            transform='rotate(17.3473 516.6 97.3374)'
-            fill='url(#paint21_linear_1_4116)'
-          />
-          <rect
-            x='349.258'
-            y='152.224'
-            width='7.04671'
-            height='8.22116'
-            rx='1'
-            transform='rotate(17.3473 349.258 152.224)'
-            fill='url(#paint22_linear_1_4116)'
-          />
-          <rect
-            x='478.215'
-            y='83.4194'
-            width='7.04671'
-            height='8.22116'
-            rx='1'
-            transform='rotate(-92.65 478.215 83.4194)'
-            fill='url(#paint23_linear_1_4116)'
-          />
-          <rect
-            x='290.641'
-            y='220.967'
-            width='10.8662'
-            height='12.6772'
-            rx='1'
-            transform='rotate(135.279 290.641 220.967)'
-            fill='url(#paint24_linear_1_4116)'
-          />
-          <rect
-            x='676.133'
-            y='378.727'
-            width='10.8662'
-            height='12.6772'
-            rx='1'
-            transform='rotate(-11.7361 676.133 378.727)'
-            fill='url(#paint25_linear_1_4116)'
-          />
-          <defs>
-            <filter
-              id='filter0_d_1_4116'
-              x='444.984'
-              y='230.809'
-              width='68.5981'
-              height='59.4336'
-              filterUnits='userSpaceOnUse'
-              color-interpolation-filters='sRGB'
-            >
-              <feFlood flood-opacity='0' result='BackgroundImageFix' />
-              <feColorMatrix
-                in='SourceAlpha'
-                type='matrix'
-                values='0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0'
-                result='hardAlpha'
-              />
-              <feOffset dy='4' />
-              <feGaussianBlur stdDeviation='1.5' />
-              <feColorMatrix
-                type='matrix'
-                values='0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.15 0'
-              />
-              <feBlend
-                mode='normal'
-                in2='BackgroundImageFix'
-                result='effect1_dropShadow_1_4116'
-              />
-              <feBlend
-                mode='normal'
-                in='SourceGraphic'
-                in2='effect1_dropShadow_1_4116'
-                result='shape'
-              />
-            </filter>
-            <linearGradient
-              id='paint0_linear_1_4116'
-              x1='492.217'
-              y1='125.337'
-              x2='492.217'
-              y2='135.023'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint1_linear_1_4116'
-              x1='622.659'
-              y1='332.32'
-              x2='622.659'
-              y2='342.006'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint2_linear_1_4116'
-              x1='389.647'
-              y1='396.5'
-              x2='389.647'
-              y2='406.186'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint3_linear_1_4116'
-              x1='434.673'
-              y1='152.546'
-              x2='434.673'
-              y2='157.015'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint4_linear_1_4116'
-              x1='616.308'
-              y1='120.925'
-              x2='616.308'
-              y2='129.156'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint5_linear_1_4116'
-              x1='603.791'
-              y1='157.066'
-              x2='603.791'
-              y2='161.535'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint6_linear_1_4116'
-              x1='592.488'
-              y1='324.957'
-              x2='592.488'
-              y2='329.426'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint7_linear_1_4116'
-              x1='623.16'
-              y1='260.384'
-              x2='623.16'
-              y2='264.853'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint8_linear_1_4116'
-              x1='666.749'
-              y1='202.268'
-              x2='666.749'
-              y2='206.737'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint9_linear_1_4116'
-              x1='631.23'
-              y1='305.584'
-              x2='631.23'
-              y2='310.054'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint10_linear_1_4116'
-              x1='552.35'
-              y1='390.856'
-              x2='552.35'
-              y2='395.325'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint11_linear_1_4116'
-              x1='391.321'
-              y1='136.435'
-              x2='391.321'
-              y2='142.816'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint12_linear_1_4116'
-              x1='287.424'
-              y1='364.433'
-              x2='287.424'
-              y2='370.814'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint13_linear_1_4116'
-              x1='372.783'
-              y1='174.824'
-              x2='372.783'
-              y2='177.241'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint14_linear_1_4116'
-              x1='584.256'
-              y1='187.74'
-              x2='584.256'
-              y2='190.156'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint15_linear_1_4116'
-              x1='669.816'
-              y1='224.868'
-              x2='669.816'
-              y2='227.285'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint16_linear_1_4116'
-              x1='389.92'
-              y1='325.549'
-              x2='389.92'
-              y2='327.966'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint17_linear_1_4116'
-              x1='314.666'
-              y1='98.9502'
-              x2='314.666'
-              y2='101.367'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint18_linear_1_4116'
-              x1='330.771'
-              y1='255.082'
-              x2='330.771'
-              y2='259.671'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint19_linear_1_4116'
-              x1='335.846'
-              y1='313.467'
-              x2='335.846'
-              y2='318.056'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint20_linear_1_4116'
-              x1='451.83'
-              y1='369.769'
-              x2='451.83'
-              y2='374.358'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint21_linear_1_4116'
-              x1='519.334'
-              y1='97.3374'
-              x2='519.334'
-              y2='103.719'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint22_linear_1_4116'
-              x1='352.782'
-              y1='152.224'
-              x2='352.782'
-              y2='160.445'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint23_linear_1_4116'
-              x1='481.739'
-              y1='83.4194'
-              x2='481.739'
-              y2='91.6406'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint24_linear_1_4116'
-              x1='296.074'
-              y1='220.967'
-              x2='296.074'
-              y2='233.645'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-            <linearGradient
-              id='paint25_linear_1_4116'
-              x1='681.566'
-              y1='378.727'
-              x2='681.566'
-              y2='391.404'
-              gradientUnits='userSpaceOnUse'
-            >
-              <stop stop-color='#020E7C' />
-              <stop offset='1' stop-color='#A3ADFF' />
-            </linearGradient>
-          </defs>
-        </svg> */}
+
         <svg
           width='167'
           height='167'
@@ -590,9 +18,9 @@ const VerificationSuccessful = () => {
             <path
               d='M54.4844 88.9547L74.8989 106.742L112.082 59.3086'
               stroke='white'
-              stroke-width='5'
-              stroke-linecap='round'
-              stroke-linejoin='round'
+              strokeWidth='5'
+              strokeLinecap='round'
+              strokeLinejoin='round'
             />
           </g>
           <defs>
@@ -605,7 +33,7 @@ const VerificationSuccessful = () => {
               filterUnits='userSpaceOnUse'
               color-interpolation-filters='sRGB'
             >
-              <feFlood flood-opacity='0' result='BackgroundImageFix' />
+              <feFlood floodOpacity='0' result='BackgroundImageFix' />
               <feColorMatrix
                 in='SourceAlpha'
                 type='matrix'
